Added Pixel Punisher
The Pixel Punisher is designed to be a semi-recreation of the Minecraft combat system in the form of a TF2 melee weapon for the Sniper. Pair it with a Huntsman for a truer Minecraft experience!

The Pixel Punisher is an experimental weapon and so is prone to a LOT of balance changes. Please give feedback by pinging me in the LFCE Discord with your thoughts for rebalancing.