Add syn_apprehension.edt to fix spawns
This EDT fixes spawns for map syn_apprehension, in the off-chance someone wants to play it (Do not do it to yourself)

However, if you DO want to play it, do the following.
Immediately after you start the map, do not choose a class, open console. Enter the following in the console:
tf_bot_quota_pvp 0; tf_bot_quota 0; mp_gamemode 6; ent_fire !self setteam 5
This will set the game to 4 team mode and place you on yellow team. Yellow is the antlion team and this map has no antlions, so you'll actually have enemies to fight. The reason for having to do this is that the vortigaunts in the map are set to team red by default, and you'd be softlocked in many areas as a result. Changing this gamemode to 4 team also makes it so you can't fail from dying.