Cyber Diver support
Adds support for Cyber Diver (put it in the sourcemods folder)

The spawn points are all messed up and many elements do not work, but the spawn points can be fixed with EDT edits down the road and the other things could be fixed too. For now, just sv_cheats 1 and noclip into the map.