Temporary fix for EP2 Outland 11
Temporary fix for a softlock in EP2 Outland 11. The scene is where you meet Kleiner and Magnusson near the rocket. The scene stops there, but it a code issue, not a map issue (From what I can tell)

For now, I have deleted the triggers and barriers for the ladder, and for the elevator you get on at the next level. This is indicated with a red light.

This version of the map will be deleted once the fix is applied for the map itself.

ep2_outland_11_original is the map that shipped with FC originally.