Update ep1_citadel_02b.edt
Replaced existing minimal EDT with a highly edited EDT from Balim's synergy repo.

This fixes issues with being able to fall without dying and causing softlocks. It also fixes the issuse with the original synergy EDT that randomly makes the platform disappear and makes everyone die near the end of the map.

This is just a bandaid fix until EDT changes are implemented better down the road.