Removed some ep1 maps, added others, add edited EDTs
Removed:
ep1_c17_00
ep1_c17_00a
ep1_c17_01
ep1_c17_01a (this map does not seem to actually be used for anything?)
Removed original LFE ep1_c17_02 (LFE changes to this map resulted in a softlock at the end of the level), but added recompiled ep1_c17_02. This is a quick dirty edit that removes the last door so you can finish the map until the coding is fixed for the sequence.

Added altered EDTs from Balim's Synergy EDT repo.

Most of these maps had multiple issues in the original versions with lighting, broken geometry, broken scripting, etc.

The EDTs added add more functionality and fix issues. Some of the EDTs have been commented out due to code causing issues with the maps and causing softlocks or other similar issues. This will have to be revisited in the future.

In ep1_c17_01, there is some commented out parts of the EDT that add additional rappellers. Though those would be very cool, it messes up the intro sequence for some reason and Alyx does not use the lines she is supposed to. They seem to rappel upon spawning in.