Multiple fixes pushed
- Fix Hitman's Heatmaker head mist effect.
This fixed the effect not working, this also solves the issue with "tf_mistify"

- Add additional NPCs for sentries to no longer target
Added Breen and G-Man to list of npcs for the sentry to ignore.
Breen is killable in various custom maps and other sourcemods and letting him be shot by sentries could sequence break in those scenarios. G-Man should always be red team, but this is being added as a pseudo-failsafe.

- Fix for "tf_disintegrate_and_gib" not working properly
Added line to code to fix effect not applying properly